+++ src/main/java/com/seoultech/ecc/admin/service/AdminMemberService.java
@@ -122,7 +122,7 @@ public class AdminMemberService {
      * 레벨 변경 요청 승인
      */
     @Transactional
-    public MemberResponse approveLevelChangeRequest(Long requestId) {
+    public MemberResponse approveLevelChangeRequest(Integer requestId) {
         LevelChangeRequestEntity request = levelChangeRequestRepository.findById(requestId)
                 .orElseThrow(() -> new RuntimeException("존재하지 않는 레벨 변경 요청입니다. ID: " + requestId));
 
@@ -146,7 +146,7 @@ public class AdminMemberService {
      * 거절 시 회원의 레벨을 변경하지 않고 요청 상태만 REJECTED로 변경
      */
     @Transactional
-    public void rejectLevelChangeRequest(Long requestId) {
+    public void rejectLevelChangeRequest(Integer requestId) {
         LevelChangeRequestEntity request = levelChangeRequestRepository.findById(requestId)
                 .orElseThrow(() -> new RuntimeException("존재하지 않는 레벨 변경 요청입니다. ID: " + requestId));
