+++ src/main/java/com/seoultech/ecc/team/controller/OneTimeTeamController.java
@@ -45,7 +45,7 @@ public class OneTimeTeamController {
     @GetMapping("/{teamId}")
     @Operation(summary = "번개 스터디 상세 정보 조회", description = "특정 번개 스터디의 상세 정보를 조회합니다.")
     public ResponseEntity<ResponseDto<OneTimeTeamDto.DetailResponse>> getOneTimeTeamDetail(
-            @PathVariable Long teamId,
+            @PathVariable Integer teamId,
             @AuthenticationPrincipal CustomUserDetails userDetails
     ) {
         Integer uuid = userDetails.getId();
@@ -83,7 +83,7 @@ public class OneTimeTeamController {
             )
     )
     public ResponseEntity<ResponseDto<OneTimeTeamDto.Response>> updateOneTimeTeam(
-            @PathVariable Long teamId,
+            @PathVariable Integer teamId,
             @AuthenticationPrincipal CustomUserDetails userDetails,
             @Valid @RequestBody OneTimeTeamDto.UpdateRequest request
     ) {
@@ -97,7 +97,7 @@ public class OneTimeTeamController {
     @PreAuthorize("hasRole('USER')")
     @Operation(summary = "번개 스터디 신청", description = "번개 스터디에 참여 신청합니다.")
     public ResponseEntity<ResponseDto<OneTimeTeamDto.Response>> applyToOneTimeTeam(
-            @PathVariable Long teamId,
+            @PathVariable Integer teamId,
             @AuthenticationPrincipal CustomUserDetails userDetails
     ) {
         Integer uuid = userDetails.getId();
@@ -110,7 +110,7 @@ public class OneTimeTeamController {
     @PreAuthorize("hasRole('USER')")
     @Operation(summary = "번개 스터디 신청 취소", description = "번개 스터디 참여 신청을 취소합니다. 스터디 시작 3시간 전까지만 취소 가능합니다.")
     public ResponseEntity<ResponseDto<OneTimeTeamDto.Response>> cancelOneTimeTeamApplication(
-            @PathVariable Long teamId,
+            @PathVariable Integer teamId,
             @AuthenticationPrincipal CustomUserDetails userDetails
     ) {
         Integer uuid = userDetails.getId();
@@ -123,7 +123,7 @@ public class OneTimeTeamController {
     @PreAuthorize("hasRole('USER')")
     @Operation(summary = "번개 스터디 취소", description = "번개 스터디를 취소합니다. 생성자만 취소 가능합니다.")
     public ResponseEntity<ResponseDto<Void>> cancelOneTimeTeam(
-            @PathVariable Long teamId,
+            @PathVariable Integer teamId,
             @AuthenticationPrincipal CustomUserDetails userDetails
     ) {
         Integer uuid = userDetails.getId();
